Factor out ID printing code into common.c
The printing output should be identical. The id structure
printing functions can be more or less used by lsnvme.
